refactoring: added jsDocs for the account type card.props and also file renaming
