+++ layouts/auth-layout.vue
@@ -2,7 +2,7 @@
 	<div class="relative flex flex-col min-h-screen pb-16 bg-primary">
 		<AuthNavbar />
 	
-      <div class="relative z-10 mt-10 md:mt-20">
+      <div class="relative z-10 md:mt-20">
          <slot />
       </div>
